pyproject.toml
[project]
name = "tech-codebase-indexer"
version = "0.1.0"
description = "AI-powered codebase indexing and semantic search system using Google ADK"
requires-python = ">=3.10"
readme = "README.md"
authors = [
    { name = "Turner Lee" }
]

dependencies = [
    # ADK and base dependencies
    "google-adk>=1.13.0",
    "python-dotenv>=1.1.0",
    
    # FastAPI and web server
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    
    # Codebase indexing
    "tree-sitter==0.25.1",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "gitpython>=3.1.45",
    "tqdm>=4.65.0",
    
    # Supabase and vector database
    "supabase>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pandas>=1.5.0",
    
    # AI and embeddings
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "codebase",
    "routers", 
    "models",
    "dev_agent",
    "tools"
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]